Move routes to separate files
