+++ server/routes/ping.js
@@ -1,0 +1,7 @@
+module.exports = {
+  method: 'GET',
+  path: '/ping',
+  handler: function(req, reply) {
+    return reply({version: require('../../package.json').version});
+  },
+};
