+++ server/routes/react-router.js
@@ -1,0 +1,8 @@
+module.exports = {
+  method: 'GET',
+  path: '/{param*}',
+  handler: function(req, reply) {
+    // use react-router routes
+    reply({message: "unimplemented"});
+  },
+};
