+++ server/routes/root.js
@@ -1,0 +1,7 @@
+module.exports = {
+  method: 'GET',
+  path: '/',
+  handler: function(req, reply) {
+    return reply.file('../build/index.html');
+  },
+};
