+++ server/routes/static.js
@@ -1,0 +1,11 @@
+module.exports = {
+  method: 'GET',
+  path: '/static/{param*}',
+  handler: {
+    directory: {
+      path: '.',
+      redirectToSlash: true,
+      index: true,
+    },
+  },
+};
